backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

